requirements.txt
decorator==4.0.10
Django==1.10.1
networkx==1.11
publicsuffix==1.1.0
PyYAML==3.12
url==0.2.0
wheel==0.24.0
